Move the spec rebuild into its own module
